pyproject.toml
[tool.poetry]
name = "weird-o"
version = "0.1.0"
description = ""
authors = ["RohanReddy44 <alletirohan@sahaj.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy-stl = "^3.2.0"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
